+++ src/Synapse/Synapse/Commands/DataPlaneCommands/Artifact/LinkConnection/UpdateAzureSynapseLinkConnectionLandingZoneCredential.cs
@@ -16,37 +16,47 @@ using Azure.Analytics.Synapse.Artifacts.Models;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Commands.Synapse.Common;
 using Microsoft.Azure.Commands.Synapse.Models;
+using Microsoft.Azure.Commands.Synapse.Properties;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
+using System;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Synapse
 {
     [Cmdlet(VerbsData.Update, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.LinkConnectionLandingZoneCredential,
-        DefaultParameterSetName = SetByName, SupportsShouldProcess = true)]
-    [OutputType(typeof(bool))]
+        DefaultParameterSetName = UpdateByName, SupportsShouldProcess = true)]
+    [OutputType(typeof(void))]
     public class UpdateAzureSynapseLinkConnectionLandingZoneCredential : SynapseArtifactsCmdletBase
     {
-        private const string SetByName = "SetByName";
-        private const string SetByObject = "SetByObject";
+        private const string UpdateByName = "UpdateByName";
+        private const string UpdateByObject = "UpdateByObject";
+        private const string UpdateByInputObject = "UpdateByInputObject";
 
-        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = SetByName,
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = UpdateByName,
             Mandatory = true, HelpMessage = HelpMessages.WorkspaceName)]
         [ResourceNameCompleter(ResourceTypes.Workspace, "ResourceGroupName")]
         [ValidateNotNullOrEmpty]
         public override string WorkspaceName { get; set; }
 
-        [Parameter(ValueFromPipeline = true, ParameterSetName = SetByObject,
+        [Parameter(ValueFromPipeline = true, ParameterSetName = UpdateByObject,
             Mandatory = true, HelpMessage = HelpMessages.WorkspaceObject)]
         [ValidateNotNull]
         public PSSynapseWorkspace WorkspaceObject { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.LinkConnectionName)]
-        [ValidateNotNullOrEmpty]
+        [Parameter(ValueFromPipeline = true, ParameterSetName = UpdateByInputObject,
+            Mandatory = true, HelpMessage = HelpMessages.LinkConnectionObject)]
+        [ValidateNotNull]
+        public PSLinkConnectionResource InputObject { get; set; }
+
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = UpdateByName,
+            Mandatory = true, HelpMessage = HelpMessages.LinkConnectionName)]
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = UpdateByObject,
+            Mandatory = true, HelpMessage = HelpMessages.LinkConnectionName)]
         public string LinkConnectionName { get; set; }
 
-        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.NotebookName)]
+        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.SasToken)]
         [ValidateNotNullOrEmpty]
-        public string SasToken { get; set; }
+        public SecureString SasToken { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -55,10 +65,19 @@ namespace Microsoft.Azure.Commands.Synapse
                 this.WorkspaceName = this.WorkspaceObject.Name;
             }
 
-            var updateLandingZoneCredential = new UpdateLandingZoneCredential();
-            updateLandingZoneCredential.SasToken = new SecureString(this.SasToken);
+            if (this.IsParameterBound(c => c.InputObject))
+            {
+                this.WorkspaceName = this.InputObject.WorkspaceName;
+                this.LinkConnectionName = this.InputObject.Name;
+            }
 
-            SynapseAnalyticsClient.UpdateLandingZoneCredential(LinkConnectionName, updateLandingZoneCredential);
+            var updateLandingZoneCredential = new UpdateLandingZoneCredential();
+            updateLandingZoneCredential.SasToken = this.SasToken;
+
+            if (this.ShouldProcess(this.WorkspaceName, String.Format(Resources.UpdatingLinkConnectionLandingZoneCredential, this.LinkConnectionName, this.WorkspaceName)))
+            {
+                SynapseAnalyticsClient.UpdateLandingZoneCredential(this.LinkConnectionName, updateLandingZoneCredential);
+            }
         }
     }
 }